+++ lib/models/layers/horizontal_band.dart
@@ -25,8 +25,8 @@ class HorizontalBand extends Layer {
         id: json['id'],
         type:
             (json['type'] as String).toLayerType() ?? LayerType.horizontalBand,
-        value: json['value'] ?? 0.0,
-        allowedError: json['allowedError'] ?? 40.0,
+        value: json['value'].toDouble() ?? 0.0,
+        allowedError: json['allowedError'].toDouble() ?? 40.0,
         color: colorFromJson(json['color']),
         isLocked: json['isLocked'] ?? false);
   }
