+++ lib/models/layers/rect_area.dart
@@ -64,9 +64,9 @@ class RectArea extends Layer {
         bottomLeft: offsetFromJson(json['bottomLeft']),
         bottomRight: offsetFromJson(json['bottomRight']),
         color: colorFromJson(json['color']),
-        strokeWidth: json['strokeWidth'] ?? 2.0,
+        strokeWidth: json['strokeWidth'].toDouble() ?? 2.0,
         alpha: json['alpha'] ?? 51,
-        endPointRadius: json['endPointRadius'] ?? 5.0,
+        endPointRadius: json['endPointRadius'].toDouble() ?? 5.0,
         isLocked: json['isLocked'] ?? false);
   }
